requirements.txt
numpy==1.15.2
pydub==0.23.0
python-telegram-bot==12.6.1
requests==2.23.0